refactor: esModule enable by default
