+++ package.json
@@ -76,7 +76,6 @@
     "file-loader": "^6.0.0",
     "husky": "^4.2.5",
     "jest": "^26.1.0",
-    "jsdom": "^16.3.0",
     "lint-staged": "^10.2.11",
     "memfs": "^3.2.0",
     "mini-css-extract-plugin": "^0.9.0",
