+++ test/fixtures/es-module/template/index.js
@@ -1,0 +1,6 @@
+import css from './index.css';
+import html from './template.js';
+
+const result = {css, html};
+
+export default  result;
