+++ test/loader.test.js
@@ -11,11 +11,7 @@ import {
   getExecutedCode,
   getModuleSource,
   getWarnings,
-  readAsset,
-  runInJsDom,
 } from './helpers/index';
-
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 jest.setTimeout(10000);
 
@@ -415,138 +411,4 @@ describe('loader', () => {
     expect(getWarnings(stats)).toMatchSnapshot('warnings');
     expect(getErrors(stats)).toMatchSnapshot('errors');
   });
-
-  const miniCssExtractPluginTests = [
-    {
-      localLoaderMode: 'commonjs',
-      extractLoaderMode: 'commonjs',
-    },
-    {
-      localLoaderMode: 'esModule',
-      extractLoaderMode: 'esModule',
-    },
-    {
-      localLoaderMode: 'commonjs',
-      extractLoaderMode: 'esModule',
-    },
-    {
-      localLoaderMode: 'esModule',
-      extractLoaderMode: 'commonjs',
-    },
-  ];
-
-  for (const test of miniCssExtractPluginTests) {
-    it(`should work with ${test.localLoaderMode}-css-loader + ${test.extractLoaderMode}-mini-css-extract-plugin`, async () => {
-      const compiler = getCompiler(
-        './simple.js',
-        {
-          esModule: test.localLoaderMode === 'esModule',
-        },
-        {
-          module: {
-            rules: [
-              {
-                test: /\.css$/i,
-                use: [
-                  {
-                    loader: MiniCssExtractPlugin.loader,
-                    options: {
-                      esModule: test.extractLoaderMode === 'esModule',
-                    },
-                  },
-                  {
-                    loader: path.resolve(__dirname, '../src'),
-                    options: {},
-                  },
-                ],
-              },
-              {
-                test: /\.(png|jpg|gif|svg|eot|ttf|woff|woff2)$/i,
-                loader: 'file-loader',
-                options: { name: '[name].[ext]' },
-              },
-            ],
-          },
-          plugins: [
-            new MiniCssExtractPlugin({
-              filename: '[name].css',
-              chunkFilename: '[id].css',
-            }),
-          ],
-        }
-      );
-      const stats = await compile(compiler);
-
-      expect(readAsset('main.css', compiler, stats)).toMatchSnapshot('css');
-      expect(
-        getExecutedCode('main.bundle.js', compiler, stats)
-      ).toMatchSnapshot('result');
-      expect(getWarnings(stats)).toMatchSnapshot('warnings');
-      expect(getErrors(stats)).toMatchSnapshot('errors');
-    });
-  }
-
-  const styleLoaderTests = [
-    {
-      localLoaderMode: 'commonjs',
-      extractLoaderMode: 'commonjs',
-    },
-    {
-      localLoaderMode: 'esModule',
-      extractLoaderMode: 'esModule',
-    },
-    {
-      localLoaderMode: 'commonjs',
-      extractLoaderMode: 'esModule',
-    },
-    {
-      localLoaderMode: 'esModule',
-      extractLoaderMode: 'commonjs',
-    },
-  ];
-
-  for (const test of styleLoaderTests) {
-    it(`should work with ${test.localLoaderMode}-css-loader + ${test.extractLoaderMode}-style-loader`, async () => {
-      const compiler = getCompiler(
-        './simple-style.js',
-        {
-          esModule: test.localLoaderMode === 'esModule',
-        },
-        {
-          module: {
-            rules: [
-              {
-                test: /\.css$/i,
-                use: [
-                  {
-                    loader: 'style-loader',
-                    options: {
-                      esModule: test.extractLoaderMode === 'esModule',
-                    },
-                  },
-                  {
-                    loader: path.resolve(__dirname, '../src'),
-                    options: {},
-                  },
-                ],
-              },
-              {
-                test: /\.(png|jpg|gif|svg|eot|ttf|woff|woff2)$/i,
-                loader: 'file-loader',
-                options: { name: '[name].[ext]' },
-              },
-            ],
-          },
-        }
-      );
-      const stats = await compile(compiler);
-
-      runInJsDom('main.bundle.js', compiler, stats, (dom) => {
-        expect(dom.serialize()).toMatchSnapshot('DOM');
-      });
-
-      expect(getWarnings(stats)).toMatchSnapshot('warnings');
-      expect(getErrors(stats)).toMatchSnapshot('errors');
-    });
-  }
 });
